updates from feedback

+++ 2013/meetup/step5_test.go
@@ -12,11 +12,16 @@ import (
 const DegToRad = math.Pi / 180
 
 func TestComputeSpiro(t *testing.T) {
-	for g := 0; g <= 360; g += 10 {
-		x, y := computeSpiro(float64(g)*DegToRad, 10, 100, 2)
+	for g := 0.0; g <= 360.0; g += 10.0 {
+		x, y := computeSpiro(g*DegToRad, 10, 100, 2)
+
+		// this time no assertions, just dump
+		//
 		fmt.Printf("(%f,%f)\n", x, y)
 	}
 }
+
+// END OMIT
 
 func computeSpiro(t float64, r float64, R float64, offset float64) (x, y float64) {
 	k := r / R
@@ -29,5 +34,3 @@ func computeSpiro(t float64, r float64, R float64, offset float64) (x, y float64
 	y = R * (y1 - y2)
 	return
 }
-
-// END OMIT
